# packages/api-client/openapi.yaml
openapi: 3.1.0
info:
  title: LitMood API
  description: 책·영화·음악을 무드와 함께 기록하고 나누는 서비스의 API
  version: v1
servers:
  - url: http://localhost:8080
    description: local
security:
  - bearerAuth: []
...
paths:
  /api/v1/users/@{handle}/follow:
    post:
      tags:
        - Social
      summary: 팔로우
      description: 비대칭 관계 — 승인이 필요 없다. 이미 팔로우 중이면 멱등하게 성공한다.
      operationId: follow
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
    delete:
      tags:
        - Social
      summary: 언팔로우
      operationId: unfollow
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/v1/users/@{handle}/block:
    post:
      tags:
        - Social
      summary: 차단
      description: 양방향으로 가려지며 기존 팔로우 관계는 양쪽 모두 해제된다
      operationId: block
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
    delete:
      tags:
        - Social
      summary: 차단 해제
      operationId: unblock
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/v1/reports:
    post:
      tags:
        - Social
      summary: 신고
      description: 같은 대상을 반복 신고해도 한 번만 접수된다
      operationId: report
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequest'
        required: true
      responses:
        '200':
          description: OK
  /api/v1/records:
    post:
      tags:
        - Records
      summary: 기록 생성
      description: 콘텐츠가 자체 DB 에 없으면 provider 에서 가져와 스냅샷을 만든다. 필수 입력은 status 뿐이다.
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecordRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RecordResponse'
  /api/v1/records/{id}/like:
    post:
      tags:
        - Social
      summary: 기록 좋아요
      operationId: likeRecord
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LikeResponse'
    delete:
      tags:
        - Social
      summary: 기록 좋아요 취소
      operationId: unlikeRecord
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LikeResponse'
...
  /api/v1/collections/{slug}/like:
    post:
      tags:
        - Social
      summary: 컬렉션 좋아요
      operationId: likeCollection
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LikeResponse'
    delete:
      tags:
        - Social
      summary: 컬렉션 좋아요 취소
      operationId: unlikeCollection
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LikeResponse'
...
  /api/v1/records/{id}:
    get:
      tags:
        - Records
      summary: 기록 단건 조회
      description: 공개 범위를 벗어나면 404 로 응답한다
      operationId: get
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RecordResponse'
    delete:
      tags:
        - Records
      summary: 기록 삭제
      description: soft delete
      operationId: delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
    patch:
      tags:
        - Records
      summary: 기록 수정
      operationId: update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRecordRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RecordResponse'
...
  /api/v1/users/@{handle}/collections:
    get:
      tags:
        - Users
      summary: 공개 컬렉션 목록
      description: 본인이 조회하면 비공개 컬렉션도 포함된다
      operationId: collections
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CollectionSummary'
      security: []
  /api/v1/records/me:
    get:
      tags:
        - Records
      summary: 내 타임라인
      description: 커서 기반 페이지네이션
      operationId: myTimeline
      parameters:
        - name: types
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
              enum:
                - BOOK
                - MOVIE
                - MUSIC
        - name: status
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
              enum:
                - WANT
                - DOING
                - DONE
                - DROPPED
        - name: moods
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
        - name: minRating
          in: query
          required: false
          schema:
            type: number
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RecordPage'
  /api/v1/records/feed:
    get:
      tags:
        - Records
      summary: 팔로잉 피드
      description: 팔로우한 사용자들의 공개·팔로워 공개 기록. 차단한 사용자는 제외된다.
      operationId: feed
      parameters:
        - name: types
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
              enum:
                - BOOK
                - MOVIE
                - MUSIC
        - name: moods
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RecordPage'
...
components:
  schemas:
    ReportRequest:
      type: object
      properties:
        targetType:
          type: string
          enum:
            - RECORD
            - COLLECTION
            - USER
        targetId:
          type: integer
          format: int64
        reason:
          type: string
          enum:
            - SPAM
            - ABUSE
            - SEXUAL
            - SPOILER
            - COPYRIGHT
            - OTHER
        detail:
          type: string
          maxLength: 500
          minLength: 0
      required:
        - reason
        - targetId
        - targetType
    CreateRecordRequest:
      type: object
      properties:
        provider:
          type: string
          enum:
            - NAVER_BOOK
            - TMDB
            - SPOTIFY
        externalId:
          type: string
        status:
          type: string
          enum:
            - WANT
            - DOING
            - DONE
            - DROPPED
        rating:
          type: number
          maximum: 5.0
          minimum: 0.5
        moods:
          type: array
          items:
            type: string
            maxLength: 30
            minLength: 0
          maxItems: 5
          minItems: 0
        review:
          type: string
          maxLength: 2000
          minLength: 0
        isSpoiler:
          type: boolean
        visibility:
          type: string
          enum:
            - PUBLIC
            - FOLLOWERS
            - PRIVATE
        contextNote:
          type: string
          maxLength: 200
          minLength: 0
        startedAt:
          type: string
          format: date
        finishedAt:
          type: string
          format: date
        repeatCount:
          type: integer
          format: int32
      required:
        - externalId
        - provider
        - status
    ContentRef:
      type: object
      description: 기록에 붙는 콘텐츠 요약
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum:
            - BOOK
            - MOVIE
            - MUSIC
        provider:
          type: string
          enum:
            - NAVER_BOOK
            - TMDB
            - SPOTIFY
        externalId:
          type: string
        title:
          type: string
        creators:
          type: array
          items:
            type: string
        releasedOn:
          type: string
          format: date
        coverUrl:
          type: string
    MoodTag:
      type: object
      properties:
        name:
          type: string
        displayName:
          type: string
        color:
          type: string
        curated:
          type: boolean
    RecordResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          type: string
          enum:
            - WANT
            - DOING
            - DONE
            - DROPPED
        rating:
          type: number
        moods:
          type: array
          items:
            $ref: '#/components/schemas/MoodTag'
        review:
          type: string
        isSpoiler:
          type: boolean
        visibility:
          type: string
          enum:
            - PUBLIC
            - FOLLOWERS
            - PRIVATE
        contextNote:
          type: string
        startedAt:
          type: string
          format: date
        finishedAt:
          type: string
          format: date
        repeatCount:
          type: integer
          format: int32
        likeCount:
          type: integer
          format: int32
        likedByMe:
          type: boolean
        authorHandle:
          type: string
        authorNickname:
          type: string
        content:
          $ref: '#/components/schemas/ContentRef'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LikeResponse:
      type: object
      properties:
        likeCount:
          type: integer
          format: int32
        likedByMe:
          type: boolean
...
    UpdateRecordRequest:
      type: object
      properties:
        status:
          type: string
          enum:
            - WANT
            - DOING
            - DONE
            - DROPPED
        rating:
          type: number
          maximum: 5.0
          minimum: 0.5
        moods:
          type: array
          items:
            type: string
            maxLength: 30
            minLength: 0
          maxItems: 5
          minItems: 0
        review:
          type: string
          maxLength: 2000
          minLength: 0
        isSpoiler:
          type: boolean
        visibility:
          type: string
          enum:
            - PUBLIC
            - FOLLOWERS
            - PRIVATE
        contextNote:
          type: string
          maxLength: 200
          minLength: 0
        startedAt:
          type: string
          format: date
        finishedAt:
          type: string
          format: date
        repeatCount:
          type: integer
          format: int32
...
    RecordPage:
      type: object
      description: 커서 기반 페이지. nextCursor 가 null 이면 마지막 페이지다.
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RecordResponse'
        nextCursor:
          type: string
        totalCount:
          type: integer
          format: int64
    CollectionSummary:
      type: object
      description: 목록용 — 아이템 없이 개수만
      properties:
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        coverUrl:
          type: string
        visibility:
          type: string
          enum:
            - PUBLIC
            - FOLLOWERS
            - PRIVATE
        itemCount:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT